+++ src/app/layout/mainlayout/mainlayout.spec.ts
@@ -1,18 +1,18 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { Dashboard } from './dashboard';
+import { Mainlayout } from './mainlayout';
 
-describe('Dashboard', () => {
-  let component: Dashboard;
-  let fixture: ComponentFixture<Dashboard>;
+describe('Mainlayout', () => {
+  let component: Mainlayout;
+  let fixture: ComponentFixture<Mainlayout>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [Dashboard]
+      imports: [Mainlayout]
     })
     .compileComponents();
 
-    fixture = TestBed.createComponent(Dashboard);
+    fixture = TestBed.createComponent(Mainlayout);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });