+++ src/app/pages/home/home.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { CardModule } from 'primeng/card';
+
+
+@Component({
+  selector: 'app-home',
+  imports: [CommonModule, CardModule],
+  templateUrl: './home.html',
+  styleUrl: './home.css',
+})
+export class Home {
+
+}